Refactor (see pull request)
