+++ Components/NavBar.tsx
@@ -1,5 +1,4 @@
-import React from 'react'
-import "./navbar.css"
+import "./Styles/navbar.css"
 
 const NavBar = () => {
   return (
